+++ crates/mimic_schema/src/node/mod.rs
@@ -1,6 +1,5 @@
 mod arg;
 mod canister;
-mod constant;
 mod def;
 mod entity;
 mod r#enum;
@@ -24,7 +23,6 @@ mod value;
 
 pub use self::arg::*;
 pub use self::canister::*;
-pub use self::constant::*;
 pub use self::def::*;
 pub use self::entity::*;
 pub use self::r#enum::*;