+++ crates/mimic_schema/src/node/schema.rs
@@ -9,7 +9,6 @@ use std::{any::Any, collections::BTreeMap};
 #[derive(Clone, Debug, Serialize)]
 pub enum SchemaNode {
     Canister(Canister),
-    Constant(Constant),
     Entity(Entity),
     Enum(Enum),
     EnumValue(EnumValue),
@@ -47,7 +46,6 @@ impl SchemaNode {
     const fn def(&self) -> &Def {
         match self {
             Self::Canister(n) => &n.def,
-            Self::Constant(n) => &n.def,
             Self::Entity(n) => &n.def,
             Self::Enum(n) => &n.def,
             Self::EnumValue(n) => &n.def,
@@ -69,7 +67,6 @@ impl MacroNode for SchemaNode {
     fn as_any(&self) -> &dyn Any {
         match self {
             Self::Canister(n) => n.as_any(),
-            Self::Constant(n) => n.as_any(),
             Self::Entity(n) => n.as_any(),
             Self::Enum(n) => n.as_any(),
             Self::EnumValue(n) => n.as_any(),
@@ -93,7 +90,6 @@ impl VisitableNode for SchemaNode {
     fn drive<V: Visitor>(&self, v: &mut V) {
         match self {
             Self::Canister(n) => n.accept(v),
-            Self::Constant(n) => n.accept(v),
             Self::Entity(n) => n.accept(v),
             Self::Enum(n) => n.accept(v),
             Self::EnumValue(n) => n.accept(v),