+++ crates/mimic_tests/design/src/test/mod.rs
@@ -33,13 +33,6 @@ pub struct Map {}
 pub struct Set {}
 
 ///
-/// Constant
-///
-
-#[constant(ty = "Str", value = "Hello")]
-pub struct CONSTANT {}
-
-///
 /// EntityIdTest
 ///